requirements.txt
boto3==1.12.26
botocore==1.15.26
docutils==0.15.2
future==0.18.2
jmespath==0.9.5
joblib==0.14.1
nltk==3.4.5
numpy==1.18.2
pandas==1.0.3
PyAthena==1.10.2
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
s3transfer==0.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
tenacity==6.1.0
urllib3==1.25.8
boto3==1.12.26
botocore==1.15.26
decorator==4.4.2
docutils==0.15.2
future==0.18.2
jmespath==0.9.5
joblib==0.14.1
networkx==2.4
nltk==3.4.5
numpy==1.18.2
pandas==1.0.3
PyAthena==1.10.2
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
s3transfer==0.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
tenacity==6.1.0
tqdm==4.43.0
urllib3==1.25.8
